tests/wpt/web-platform-tests/old-tests/submission/Opera/script_scheduling/092.html

<!DOCTYPE html>
<html><head>
	<title> scheduler: defer script and slow-loading non-async external script</title>
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="testlib/testlib.js"></script>
</head>
<body>

	<div id="log">FAILED (This TC requires JavaScript enabled)</div>

	<script>
          var t = async_test(undefined, {timeout:3500});
          
          var script = document.createElement("script");
          script.src = "scripts/include-2.js?pipe=trickle(d2)";
          script.async = false;
          document.head.appendChild(script);

	  onload = t.step_func(function() {
                                 assert_array_equals(eventOrder, ['external script #1', 'external script #2']);
                                 t.done();
                             });
        </script>
        <script defer src="scripts/include-1.js"></script>
</body></html>
